requirements.txt
flask>=3.0
flask-socketio>=5.3
python-engineio>=4.0
python-socketio>=5.0
pandas>=1.4.3,<3
telethon==1.42.0
requests>=2.28.1
websockets>=12.0.1
pyOpenSSL>=20.0.1
python-dotenv>=1.0
